# internal/openapi/openapi.yaml
openapi: "3.0.3"
info:
  title: FabAID Manager API
  description: NSF grant/project tracking — effort, budgets, subawards, invoices, and statements of work.
  version: "1.0.0"
  license:
    name: Apache-2.0
servers:
  - url: /api/v1
    description: API v1
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Session cookie or API key (Bearer fabaid_…)
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
...
    APIKey:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
        key_prefix:
          type: string
          readOnly: true
        roles:
          type: array
          items:
            type: string
        created_by:
          type: string
        created_by_name:
          type: string
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        last_used_at:
          type: string
          format: date-time
          nullable: true
          readOnly: true
        idle_timeout_s:
          type: integer
          nullable: true
        expires_at:
          type: string
          format: date-time
          nullable: true
        revoked_at:
          type: string
          format: date-time
          nullable: true
          readOnly: true
    APIKeyCreateResponse:
      allOf:
        - $ref: "#/components/schemas/APIKey"
        - type: object
          properties:
            raw_key:
              type: string
              description: Shown exactly once at creation time.
...
    BackupSettings:
      type: object
      properties:
        backup_frequency_hours:
          type: integer
        backup_bucket:
          type: string
        backup_endpoint:
          type: string
        backup_access_key:
          type: string
        backup_secret_key:
          type: string
        backup_use_ssl:
          type: boolean
...
security:
  - bearerAuth: []
  - cookieAuth: []
paths:
...
  /admin/api-keys:
    get:
      tags: [admin]
      summary: List API keys
      operationId: listAPIKeys
      responses:
        "200":
          description: Array of API keys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/APIKey"
    post:
      tags: [admin]
      summary: Create an API key
      operationId: createAPIKey
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                roles:
                  type: array
                  items:
                    type: string
      responses:
        "201":
          description: Created API key (includes plain-text key — shown only once)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIKeyCreateResponse"
  /admin/api-keys/{keyID}/revoke:
    parameters:
      - name: keyID
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [admin]
      summary: Revoke an API key
      operationId: revokeAPIKey
      responses:
        "200":
          description: API key revoked
  /admin/api-keys/{keyID}:
    parameters:
      - name: keyID
        in: path
        required: true
        schema:
          type: string
          format: uuid
    delete:
      tags: [admin]
      summary: Delete an API key
      operationId: deleteAPIKey
      responses:
        "204":
          description: Deleted

  # ── Backups ────────────────────────────────────────────
...
  /backups/upload-restore:
    post:
      tags: [backups]
      summary: Upload and restore from a backup file (admin)
      operationId: uploadRestore
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                encryption_key:
                  type: string
      responses:
        "200":
          description: Restore triggered
  /backups/settings:
    get:
      tags: [backups]
      summary: Get backup settings (admin)
      operationId: getBackupSettings
      responses:
        "200":
          description: Backup settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BackupSettings"
    put:
      tags: [backups]
      summary: Update backup settings (admin)
      operationId: updateBackupSettings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BackupSettings"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BackupSettings"
...
  /backups/derive-key:
    post:
      tags: [backups]
      summary: Derive an encryption key from a passphrase (admin)
      operationId: deriveKeyFromInput
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [passphrase]
              properties:
                passphrase:
                  type: string
      responses:
        "200":
          description: Derived key
          content:
            application/json:
              schema:
                type: object
                properties:
                  key:
                    type: string
  /backups/{backupID}/download:
    parameters:
      - name: backupID
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [backups]
      summary: Download a backup archive (admin)
      operationId: downloadBackup
      responses:
        "200":
          description: Backup archive
          content:
            application/gzip:
              schema:
                type: string
                format: binary
  /backups/{backupID}/key:
    parameters:
      - name: backupID
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [backups]
      summary: Get per-backup encryption key (admin)
      operationId: getPerBackupKey
      responses:
        "200":
          description: Per-backup key
          content:
            application/json:
              schema:
                type: object
                properties:
                  key:
                    type: string
  /backups/{backupID}/restore:
    parameters:
      - name: backupID
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [backups]
      summary: Restore from an existing backup (admin)
      operationId: restoreBackup
      responses:
        "200":
          description: Restore triggered
  /backups/{backupID}:
    parameters:
      - name: backupID
        in: path
        required: true
        schema:
          type: string
          format: uuid
    delete:
      tags: [backups]
      summary: Delete a backup (admin)
      operationId: deleteBackup
      responses:
        "204":
          description: Deleted

  # ── Health Check ────────────────────────────────────────
  # NOTE: /healthz is served at the root, not under /api/v1.
...